requirements.txt
djangorestframework==3.14.0
PyJWT==2.6.0
django-cors-headers==3.14.0
# psycopg2==2.9.5
psycopg2-binary==2.9.5
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2